+++ packages/terra-clinical-header/src/terra-dev-site/doc/example/HeaderLongTextWithContent.jsx
@@ -6,8 +6,9 @@ const LongTextWithContent = () => (
   <div>
     <Header
       startContent={<Button text="Button" />}
-      title="Lorem ipsum dolor sit amet, sea meis iisque petentium an. Vero copiosae eu eos. Usu ocurreret definiebas eu, vim esse duis habemus at, vix at alii ubique. Nobis meliore principes te cum. Ei quando eirmod oportere eum, ei nonumy eruditi eum. In qui odio euripidis, habeo mundi praesent vim et."
+      text="Lorem ipsum dolor sit amet, sea meis iisque petentium an. Vero copiosae eu eos. Usu ocurreret definiebas eu, vim esse duis habemus at, vix at alii ubique. Nobis meliore principes te cum. Ei quando eirmod oportere eum, ei nonumy eruditi eum. In qui odio euripidis, habeo mundi praesent vim et."
       endContent={<Button text="Button" />}
+      level={3}
     >
       <Button text="Custom 1" />
     </Header>