+++ packages/terra-clinical-header/src/terra-dev-site/doc/example/TitleHeader.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Header from 'terra-clinical-header';
 
 const TitleHeader = () => (
-  <Header title="Default Header" />
+  <Header text="Default Header" level={5} />
 );
 
 export default TitleHeader;