+++ packages/terra-clinical-header/src/terra-dev-site/test/clinical-header/DefaultHeader.test.jsx
@@ -2,4 +2,8 @@ import React from 'react';
 
 import Header from '../../../Header';
 
-export default () => <Header id="Header" />;
+export default () => (
+  <div aria-labelledby="Header">
+    <Header text="Test-Header" id="Header" level={1} />
+  </div>
+);