+++ packages/terra-clinical-header/src/terra-dev-site/test/clinical-header/NoTitleRightContentHeader.test.jsx
@@ -16,5 +16,6 @@ const content = (
 export default () => (
   <Header
     endContent={content}
+    level={1}
   />
 );