+++ packages/terra-clinical-header/src/terra-dev-site/test/clinical-header/RightContentHeader.test.jsx
@@ -15,7 +15,8 @@ const content = (
 
 export default () => (
   <Header
-    title="Header with content on the right"
+    text="Header with content on the right"
     endContent={content}
+    level={1}
   />
 );